Prevent internal tick drift over time
Instead of requesting a new timestamp after each operation, we compute it from the beginning of the emulation.
This allows the emulation to wait if it is somehow faster than the real hardware at some points, and to catch up if it is slower, thus feeling accurate even if it is not at the CPU operation level.